Functioning scraper for BA!
Fueled by Hans Zimmer and weird beer. This grabs alla the
reviews from BA (seriously, all of them) and parses them into a nice
format. Currently it just prints out the first two pages of beers, but
at some point we should be putting this into a database or something.